+++ blocklets/image-bin/src/libs/constants.js
@@ -3,7 +3,7 @@ const WELL_KNOWN_SERVICE_PATH = '/.well-known/service';
 const RESOURCE_TYPE = 'imgpack';
 
 export const COMPONENT_DID = DID;
-export const PROJECT_PAGE_PATH = `${WELL_KNOWN_SERVICE_PATH}/embed/resources/${DID}/publish`;
-export const ADD_RESOURCE_PAGE_PATH = `${WELL_KNOWN_SERVICE_PATH}/embed/resources/${DID}/add?resourceType=${RESOURCE_TYPE}`;
+export const PROJECT_PAGE_PATH = `${WELL_KNOWN_SERVICE_PATH}/embed/resources/${DID}/publish?mode=dialog`;
+export const ADD_RESOURCE_PAGE_PATH = `${WELL_KNOWN_SERVICE_PATH}/embed/resources/${DID}/add?resourceType=${RESOURCE_TYPE}&mode=dialog`;
 export const RESOURCE_API_PATH = `${WELL_KNOWN_SERVICE_PATH}/api/resources?resourceType=${RESOURCE_TYPE}`;
 export { RESOURCE_TYPE };